api/requirements.txt
# FastAPI backend — minimal deps for Render deployment
# NOTE: Use `pip-compile` (from pip-tools) to generate a fully pinned
# requirements.lock for reproducible deploys.  The bounds below prevent
# accidental major-version jumps while still allowing patch updates.
fastapi>=0.129.0,<1.0
uvicorn[standard]>=0.41.0,<1.0
asyncpg>=0.30.0,<1.0
python-dotenv>=1.0.1,<2.0
pydantic>=2.10.0,<3.0
httpx>=0.28.1,<1.0
aiosmtplib>=3.0.0,<4.0
slowapi>=0.1.9,<1.0
alembic>=1.14.0,<2.0
sqlalchemy>=2.0.0,<3.0
sentry-sdk[fastapi]>=2.19.0,<3.0
bcrypt>=4.2.0,<5.0
PyJWT>=2.10.0,<3.0
openai>=1.0.0,<2.0
langfuse>=2.0.0,<3.0
PyPDF2>=3.0.0,<4.0
PyMuPDF>=1.24.0,<2.0
python-multipart>=0.0.6,<1.0
